+++ src/prefix.rs
@@ -4,12 +4,12 @@
 pub mod metric {
     pub mod f64 {
         //! Constants for metric prefixes, f64 precision
-        
-        /// 10<sup>30</sup> 
+
+        /// 10<sup>30</sup>
         pub const QUETTA: f64 = 1.0_E30;
-        /// 10<sup>27</sup> 
+        /// 10<sup>27</sup>
         pub const RONNA: f64 = 1.0_E27;
-        /// 10<sup>24</sup> 
+        /// 10<sup>24</sup>
         pub const YOTTA: f64 = 1.0_E24;
         /// 10<sup>21</sup>
         pub const ZETTA: f64 = 1.0_E21;
@@ -57,9 +57,9 @@ pub mod metric {
     }
     pub mod f32 {
         //! Constants for metric prefixes, f32 precision
-        /// 10<sup>30</sup> 
+        /// 10<sup>30</sup>
         pub const QUETTA: f32 = 1.0_E30;
-        /// 10<sup>27</sup> 
+        /// 10<sup>27</sup>
         pub const RONNA: f32 = 1.0_E27;
         /// 10<sup>24</sup>
         pub const YOTTA: f32 = 1.0_E24;