+++ src/si/extended/f32/mod.rs
@@ -1,6 +1,5 @@
 //! Quantities with f32 underlying storage type.
 
-
 mod quantity;
 pub use quantity::Quantity;
 pub mod constants;