+++ src/si/extended/f64/mod.rs
@@ -1,6 +1,5 @@
 //! Quantities with f64 underlying storage type.
 
-
 pub mod quantity;
 pub use quantity::Quantity;
 pub mod constants;