+++ src/si/isq/f32/mod.rs
@@ -1,6 +1,5 @@
 //! Quantities with 32 underlying storage type.
 
-
 mod quantity;
 pub use quantity::Quantity;
 pub mod constants;