+++ src/si/isq/f64/mod.rs
@@ -1,7 +1,6 @@
 //! Quantities with f64 underlying storage type.
 //!
 
-
 pub mod quantity;
 pub use quantity::Quantity;
 pub mod constants;